backend/requirements.txt
asgiref==3.5.2
Babel==2.11.0
black==22.12.0
click==8.1.3
Django==3.2
django-environ==0.9.0
django-extra-views==0.13.0
django-haystack==3.2.1
django-oscar==3.1
django-phonenumber-field==5.2.0
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.12
factory-boy==3.2.1
Faker==15.3.4
flake8==6.0.0
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.3
phonenumbers==8.13.2
Pillow==9.3.0
platformdirs==2.6.0
psycopg2==2.9.5
purl==1.6
pycodestyle==2.10.0
pycountry==22.3.5
pyflakes==3.0.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sorl-thumbnail==12.6.3
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
